requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
torchtext
spacy
opencv-python
natsort
sdv
